+++ src/main/java/org/jabref/gui/EntryTypeView.java
@@ -7,7 +7,9 @@ import java.util.Optional;
 import javafx.application.Platform;
 import javafx.event.Event;
 import javafx.fxml.FXML;
+
 import javax.inject.Inject;
+
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
@@ -27,6 +29,7 @@ import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.BibEntryType;
 import org.jabref.model.entry.types.BiblatexEntryTypeDefinitions;
+import org.jabref.model.entry.types.BiblatexSoftwareEntryTypeDefinitions;
 import org.jabref.model.entry.types.BibtexEntryTypeDefinitions;
 import org.jabref.model.entry.types.EntryType;
 import org.jabref.model.entry.types.IEEETranEntryTypeDefinitions;
@@ -52,10 +55,12 @@ public class EntryTypeView extends BaseDialog<EntryType> {
     @FXML private FlowPane bibTexPane;
     @FXML private FlowPane ieeetranPane;
     @FXML private FlowPane customPane;
+    @FXML private FlowPane biblatexSoftwarePane;
     @FXML private TitledPane biblatexTitlePane;
     @FXML private TitledPane bibTexTitlePane;
     @FXML private TitledPane ieeeTranTitlePane;
     @FXML private TitledPane customTitlePane;
+    @FXML private TitledPane biblatexSoftwareTitlePane;
 
     @Inject StateManager stateManager;
 
@@ -140,9 +145,11 @@ public class EntryTypeView extends BaseDialog<EntryType> {
         ieeeTranTitlePane.managedProperty().bind(ieeeTranTitlePane.visibleProperty());
         biblatexTitlePane.managedProperty().bind(biblatexTitlePane.visibleProperty());
         customTitlePane.managedProperty().bind(customTitlePane.visibleProperty());
+        biblatexSoftwareTitlePane.managedProperty().bind(biblatexSoftwareTitlePane.visibleProperty());
 
         if (basePanel.getBibDatabaseContext().isBiblatexMode()) {
             addEntriesToPane(biblatexPane, BiblatexEntryTypeDefinitions.ALL);
+            addEntriesToPane(biblatexSoftwarePane, BiblatexSoftwareEntryTypeDefinitions.ALL);
 
             bibTexTitlePane.setVisible(false);
             ieeeTranTitlePane.setVisible(false);
@@ -155,6 +162,7 @@ public class EntryTypeView extends BaseDialog<EntryType> {
             }
         } else {
             biblatexTitlePane.setVisible(false);
+            biblatexSoftwareTitlePane.setVisible(false);
             addEntriesToPane(bibTexPane, BibtexEntryTypeDefinitions.ALL);
             addEntriesToPane(ieeetranPane, IEEETranEntryTypeDefinitions.ALL);